Emergency App/backend/requirements.txt
apturl==0.5.2
asgiref==3.11.0
astroid==3.0.3
asttokens==3.0.1
attrs==21.2.0
autopep8==2.0.4
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==23.11.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.4
click==8.0.3
colorama==0.4.4
command-not-found==0.3
coverage==7.13.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
dill==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.26.5
duplicity==0.8.21
evdev==1.4.0
exceptiongroup==1.3.1
executing==2.2.1
factory-boy==3.3.0
Faker==20.1.0
fasteners==0.14.1
flake8==6.1.0
future==0.18.2
gunicorn==21.2.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.3.0
ipdb==0.13.13
ipython==8.17.2
isort==5.12.0
jedi==0.19.2
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.2.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
mypy_extensions==1.1.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==25.0
paramiko==2.9.3
parso==0.8.5
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycodestyle==2.11.1
pycups==2.0.1
pyflakes==3.1.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.8.0
pylint==3.0.3
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
python-magic==0.4.24
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
sentry-sdk==1.38.0
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.4
stack-data==0.6.3
systemd-python==234
texttable==1.6.4
tomli==2.3.0
tomlkit==0.13.3
traitlets==5.14.3
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.1.0
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.2.3
whitenoise==6.6.0
xdg==5
xkit==0.0.0
yamllint==1.26.3
zipp==1.0.0
